SDK-1470: Reduce module nesting
